First pass form submission actions. Also modularizing state structure.
