+++ kolibri/plugins/setup_wizard/assets/src/state/actions/main.js
